Reorganized the componets in finder
