+++ src/pages/show-actor/ActorDetails.jsx
@@ -1,5 +1,5 @@
 import styled from "styled-components"
-import { Media } from "./Media"
+import { Media } from "../../components/themes/Media"
 
 const ActorDetails2 = styled.div`
 display: flex;
